Changes made for Maker Faire Demo
The main reason to branch this is because of the changes to
CALIB_MAX_CHANGE and CALIB_NUM_REPEATS, which are still untested, but
made things work better for the demo. There are also some added Serial
print statements to debug. Let's merge the branches once we're done
debugging stuff.