Add while(!Serial) for reliable Serial init
If you don't include `while(!Serial);`, Clyde's serial
port is very very unreliable.  This is due to the fact
that Clyde is based on the Ardunino Leonardo, which shares
USB with serial communication.

Add a Serial output so people know that Clyde is now ready
and willing to accept commands.